src/requirements.txt
elasticsearch==2.1.0
python-nmap==0.4.4
urllib3==1.12
wheel==0.24.0
